Extract Grape::Endpoint.before_each into Grape::Testing module
Move before_each, reset_before_each, and run_before_each out of
Grape::Endpoint and into an opt-in Grape::Testing module loaded via
`require 'grape/testing'`. Production requests no longer pay any cost
from this testing-only feature. RunBeforeEach is prepended into
Endpoint#run and ClassMethods is extended onto Grape::Endpoint only
when the testing module is required.
